+++ .env.sample
@@ -7,3 +7,4 @@ MYSQL_PORT=3306
 
 BACKEND_CONTAINER_NAME=20211105_go_rest_server
 BACKEND_PORT=4000
+BACKEND_CONTAINER_PORT=4000
